+++ src/main/java/com/hubspot/jinjava/lib/tag/MacroTag.java
@@ -7,8 +7,8 @@ import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
 
-import com.google.common.base.Objects;
 import com.google.common.base.Splitter;
+import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.hubspot.jinjava.doc.annotations.JinjavaDoc;
 import com.hubspot.jinjava.doc.annotations.JinjavaParam;
@@ -70,7 +70,7 @@ public class MacroTag implements Tag {
     }
 
     String name = matcher.group(1);
-    String args = Objects.firstNonNull(matcher.group(2), "");
+    String args = Strings.nullToEmpty(matcher.group(2));
 
     LinkedHashMap<String, Object> argNamesWithDefaults = new LinkedHashMap<>();