+++ src/main/java/com/hubspot/jinjava/tree/output/OutputNode.java
@@ -4,4 +4,6 @@ public interface OutputNode {
 
   String getValue();
 
+  long getSize();
+
 }