+++ src/main/java/com/hubspot/jinjava/tree/output/RenderedOutputNode.java
@@ -1,5 +1,9 @@
 package com.hubspot.jinjava.tree.output;
 
+import java.nio.charset.Charset;
+
+import com.google.common.base.Charsets;
+
 public class RenderedOutputNode implements OutputNode {
 
   private final String output;
@@ -18,4 +22,8 @@ public class RenderedOutputNode implements OutputNode {
     return getValue();
   }
 
+  @Override
+  public long getSize() {
+    return output == null ? 0 : output.getBytes(Charset.forName(Charsets.UTF_8.name())).length;
+  }
 }